+++ dphonebook/lib/providers/__init__.py
@@ -3,6 +3,7 @@ from dphonebook.lib.providers.receive_sms_co import ReceiveSmsCo
 from dphonebook.lib.providers.receive_sms_online_info import ReceiveSmsOnlineInfo
 from dphonebook.lib.providers.receive_sms_org import ReceiveSmsOrg
 from dphonebook.lib.providers.receive_smss import ReceiveSmss
+from dphonebook.lib.providers.sms_online_co import SmsOnlineCo
 
 
 number_provider_classes = [
@@ -10,5 +11,6 @@ number_provider_classes = [
     ReceiveSmss,
     ReceiveSmsCo,
     ReceiveSmsOnlineInfo,
-    ReceiveSmsOrg
+    ReceiveSmsOrg,
+    SmsOnlineCo
 ]